backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.7
pytest==7.4.4
hypothesis==6.98.3
aiosqlite==0.19.0
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.32.5
